Updating the instantiation of the notifier
When creating an instance of the slack notifier, a parameter for
the team was being used. That is no longer needd, this commit
updates that.

Also, when I first installed the project it didn't work, cause
it installed slack-notifier gem version 0.5, which is very old,
this patch fixes the gem version to be the latest.